fix: Delegated create folder response return parse error
