+++ .changeset/nasty-tomatoes-occur.md
@@ -1,0 +1,5 @@
+---
+'@bnb-chain/greenfield-js-sdk': patch
+---
+
+fix: Delegated Create Folder APi return parse error message
