+++ examples/nextjs/src/components/object/create/index.tsx
@@ -351,6 +351,8 @@ export const CreateObject = () => {
               },
             );
 
+            console.log('res', res);
+
             if (res.code === 0) {
               alert('success');
             }
