+++ packages/js-sdk/src/types/sp/DelegateCreateFolder.ts
@@ -1,0 +1,15 @@
+import { DelegatedOpts } from './Common';
+
+export type DelegatedCreateFolderRequest = {
+  bucketName: string;
+  objectName: string;
+  delegatedOpts: DelegatedOpts;
+  endpoint?: string;
+  timeout?: number;
+};
+
+export interface DelegateCreateFolderRepsonse {
+  CreateHash: {
+    TxHash: string;
+  };
+}
