+++ packages/js-sdk/src/types/sp/DelegatedPubObject.ts
@@ -1,0 +1,12 @@
+import { DelegatedOpts, ResumableOpts } from './Common';
+
+export type DelegatedPubObjectRequest = {
+  bucketName: string;
+  objectName: string;
+  body: File;
+  delegatedOpts: DelegatedOpts;
+  endpoint?: string;
+  timeout?: number;
+  contentType?: string;
+  resumableOpts?: ResumableOpts;
+};
