+++ packages/js-sdk/src/types/sp/PutObject.ts
@@ -1,4 +1,4 @@
-import { VisibilityType } from '../common';
+import { ResumableOpts } from './Common';
 
 export type PutObjectRequest = {
   bucketName: string;
@@ -8,33 +8,4 @@ export type PutObjectRequest = {
   duration?: number;
   endpoint?: string;
   resumableOpts?: ResumableOpts;
-};
-
-export type DelegatedPubObjectRequest = {
-  bucketName: string;
-  objectName: string;
-  body: File;
-  delegatedOpts: DelegatedOpts;
-  endpoint?: string;
-  timeout?: number;
-  contentType?: string;
-  resumableOpts?: ResumableOpts;
-};
-
-export type DelegatedCreateFolderRequest = {
-  bucketName: string;
-  objectName: string;
-  delegatedOpts: DelegatedOpts;
-  endpoint?: string;
-  timeout?: number;
-};
-
-export type DelegatedOpts = {
-  visibility: VisibilityType;
-  isUpdate?: boolean;
-};
-
-export type ResumableOpts = {
-  disableResumable: boolean;
-  partSize?: number;
 };
